labeler/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pillow>=10.0.0
